extracted logic from handleConfig into corresponding handlers

+++ src/config/deleteConfig.ts
@@ -1,12 +1,29 @@
 import * as fs from "fs";
+import chalk from "chalk";
 import { getFileName, fileExists } from "../helpers/config";
-import { printSuccess } from "../helpers/print";
+import { cliPrompts } from "../helpers/prompts";
+import { printSuccess, printError } from "../helpers/print";
+import getConfigArgs from "../command/getConfigArgs";
 
 const configFileName = getFileName("config");
+const { force } = getConfigArgs();
+const { toggle } = cliPrompts;
 
-export default function deleteConfig() {
+export default async function deleteConfig() {
   if (configFileName != null && fileExists("config")) {
-    fs.unlinkSync(configFileName);
-    printSuccess("Deleted!");
+    let continueToDelete = force;
+    if (!continueToDelete) {
+      continueToDelete = await toggle(
+        `${chalk.yellow("Are you sure you want to delete the config file?\n")}`,
+        false
+      );
+    }
+
+    if (continueToDelete) {
+      fs.unlinkSync(configFileName);
+      printSuccess("Deleted!\n");
+    }
+  } else {
+    printError("There is no config file to delete.\n");
   }
 }
